% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleChart.R
\name{BubbleChart}
\alias{BubbleChart}
\title{Bubble Chart}
\usage{
BubbleChart(dat_list, fit = NULL, xlimits = c(1.5, 10.5), xbreaks = 2:10,
  plot = TRUE, cols = 2)
}
\arguments{
\item{dat_list}{a list like the one returned by \code{FormatTiters}}

\item{fit}{what type of fit to add. Current options are "lm" for linear model, "exp" for exponential, or \code{NULL} for no smoothing.}

\item{xlimits}{the x-axis limits (passed to scale_x_continuous)}

\item{xbreaks}{the x-axis breaks (passed to scale_x_continuous)}

\item{plot}{logical indicating whether to plot or not. Default is TRUE}

\item{cols}{numeric specifying how many columns to layout plot}

\item{scale_y}{a character string specifying whether the y axis should be "fixed" for all strains or "free".}
}
\value{
a list of ggplot2 objects.
}
\description{
\code{BubbleChart} visualizes baseline vs fold change in titers
}
\details{
This plot was designed for HAI titer data with baseline columns and fold change columns for multiple strains.
}
\examples{
\dontrun{
## Example using the master phenotype file
library(dplyr)
titers <- master \%>\%
  filter(Year == 1, AgeGroup \%in\% "Young", !is.na(whoResp))
titer_list <- FormatTiters(titers,
                           strains = c("A_California_7_2009",
                               "A_Perth_16_2009",
                               "B_Brisbane_60_2008"))
}
BubbleChart(titer_list)
}
\author{
Stefan Avey
}
\seealso{
\code{FormatTiters}
}
\keyword{HIPC}

