% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_padjMFC.R
\name{Calculate_D0NormPaired}
\alias{Calculate_D0NormPaired}
\title{Calculate_D0NormPaired}
\usage{
Calculate_D0NormPaired(
  dat,
  fcStdCols = grep("fc_std_norm", colnames(dat), value = TRUE)
)
}
\arguments{
\item{dat}{data frame containing \code{fcStdCols}}

\item{fcStdCols}{column names containing the titer fold changes for each strain
standardized across subjects. Length must be > 1}
}
\value{
a numeric vector containing the values from d0StdCols that
                 correspond to the maximum over the strains of fcStdCols
}
\description{
\code{Calculate_D0NormPaired} calculates the normalized day 0 titer paired with the titer with maximum normalized fold change
}
\details{
If there are multiple strains that have the maximal fold change,
choose the day 0 titer that is higher since this will allow
for a greater adjustment and better chance of being a high responder.

Column names containing the day 0 titers for each strain
standardized across subjects are assumed to follow the same
pattern as \code{fcStdCols} with "d0" replacing "fc" in the name.
}
\examples{
## First Example

}
\author{
Stefan Avey
}
