% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_ffalts.R
\name{Calculate_ffalts}
\alias{Calculate_ffalts}
\title{Calculate ffalts}
\usage{
Calculate_ffalts(dat_list, subjectCol = "SubjectID",
  responseLabels = c("NR", "X", "R"))
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{responseLabels}{names for low, middle, and high responses}
}
\value{
A named list with 1 element named "ffalts" containing the response ("NR", "X", or "R").
}
\description{
\code{Calculate_ffalts} calculates a response definition based on a Four Fold change to At Least Two Strains (ffalts).
}
\details{
Subjects are responders (default "R") if they acheive a 4-fold or greater fold change in
titer to at least 2 strains, nonresponders (default "NR") if they do not acheive a 4-fold
or greater fold change in titer to any strain, and intermediate (default "X") otherwise.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

Calculate_ffalts(titer_list)
}
\author{
Stefan Avey
}

