% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_MFC.R
\name{Calculate_MFC}
\alias{Calculate_MFC}
\title{Calculate MFC}
\usage{
Calculate_MFC(dat_list, subjectCol = "SubjectID", discretize = c(0.2, 0.3),
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"))
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{discretize}{a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20\% and 30\%.}

\item{responseLabels}{names for low, moderate and high responses}
}
\value{
A list with the following elements:
\describe{
  \item{MFC}{a named vector containing the continuous MFC endpoints}
  \item{MFC_d<X>}{a named vector containing the discrete MFC endpoint with a cutoff at <X>}
  \item{...}{Other named vectors containing discrete MFC endpoints}
}

A named list containing the MFC for each subject and any discretized metrics
}
\description{
\code{Calculate_MFC} calculates the (log-transformed) maximum fold change over all strains.
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

Calculate_MFC(titer_list)
}
\author{
Stefan Avey
}

