% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Nakaya2015.R
\name{Calculate_Nakaya2015}
\alias{Calculate_Nakaya2015}
\title{Calculate Nakaya2015}
\usage{
Calculate_Nakaya2015(dat_list, subjectCol = "SubjectID",
  responseLabels = paste0(c("low", "high"), "Responder"),
  na_action = "na.fail", ...)
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{responseLabels}{names for low and high responses}

\item{na_action}{how should missing \code{NA} values be treated. Default is "na.fail"}

\item{...}{Additional arguments passed to \code{lm}}
}
\value{
A list with the following elements:
\describe{
  \item{data}{a data frame containing the MFC and indicator variables that determine whether subject is a low or high responder (see details)}
  \item{Nakaya2015}{a named vector containing the discretized endpoint}
}
}
\description{
\code{Calculate_Nakaya2015} calculates the endpoint used in Nakaya et al. 2015
}
\details{
First calculate the maximum fold change (MFC) derived titer metric described
in Nakaya et al. 2015. Then check whether both of these conditions are satisfied:
    i) MFC is at least a 4-fold increase
   ii) The "Post" antibody titer is 1:40 or more for at least 1 strain
Subjects are classified as high responders if they satisfy both conditions and
low responders otherwise.

Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

## Calculate the endpoint
endpoints <- Calculate_Nakaya2015(titer_list)
summary(endpoints)
}
\references{
Nakaya HI, et al. (2015) Systems Analysis of Immunity to Influenza Vaccination across Multiple Years and in Diverse Populations Reveals Shared Molecular Signatures. Immunity 43(6):1186-1198.
}
\seealso{
\code{CalculateMFC}
}
\author{
Stefan Avey
}
