% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_padjMFC.R
\name{Calculate_padjMFC}
\alias{Calculate_padjMFC}
\title{Calculate_padjMFC}
\usage{
Calculate_padjMFC(
  dat,
  fcCol = "fc_norm_max_ivt",
  d0Col = "d0_norm_paired",
  discretize = c(0.2, 0.3),
  scaleResiduals = FALSE,
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"),
  ...
)
}
\arguments{
\item{dat}{the data containing the columns \code{fcCol} and \code{d0Col}}

\item{fcCol}{character string specifying the name of the fold change column from \code{dat}}

\item{d0Col}{character string specifying the name of the day 0 column from \code{dat}}

\item{discretize}{a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20\% and 30\%.}

\item{scaleResiduals}{Logical. Should residuals be scaled inversely by the
square of the confidence intervals from the linear model.}

\item{responseLabels}{names for low, moderate and high responses}

\item{...}{Additional arguments passed to \code{lm}}
}
\value{
A list with the first element named "linearModel" for the linear model and then "padjMFC" containing the continuous padjMFC metric and one additional element for each value of discretize giving the discrete labels.
}
\description{
\code{Calculate_padjMFC} calculates the paired, adjusted maximum fold change (padjMFC)
}
\details{
Calculate the paired, adjusted maximum fold change (padjMFC) from
fc_norm_max_ivt and d0_norm_paired using linear regression to
remove the effect of baseline titers. Missing (\code{NA}) values are handled
and any missing values in fcCol and d0Col will also be missing in the output.
}
\examples{
## First Example

}
\seealso{
\code{lm}
}
\author{
Stefan Avey
}
