% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_preGMT.R
\name{Calculate_preGMT}
\alias{Calculate_preGMT}
\title{Calculate pre-GMT}
\usage{
Calculate_preGMT(dat_list, subjectCol = "SubjectID")
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}
}
\value{
A named vector containing the pre-GMT for each subject
}
\description{
\code{Calculate_preGMT} calculates the log-transformed pre-vaccination geometric mean titer (pre-GMT)
}
\details{
Non-logged HAI titers for each strain are used to calculate the geometric mean and
the geometric mean for each subject is subsequently log2-transformed.
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

Calculate_preGMT(titer_list)
}
\author{
Stefan Avey
}
