% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_Thakar2015.R
\name{Calculate_Thakar2015}
\alias{Calculate_Thakar2015}
\title{Calculate Thakar2015}
\usage{
Calculate_Thakar2015(dat_list, subjectCol = "SubjectID",
  responseLabels = c("NR", "X", "R"))
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{responseLabels}{names for low, middle, and high responses}
}
\value{
A named list with 1 element named "Thakar2015" containing the response ("NR", "X", or "R").
}
\description{
\code{Calculate_Thakar2015} calculates a response definition based on a Four Fold change to At Least Two Strains (Thakar2015).
}
\details{
Subjects are responders (default "R") if they acheive a 4-fold or greater fold change in
titer to at least 2 strains, nonresponders (default "NR") if they do not acheive a 4-fold
or greater fold change in titer to any strain, and intermediate (default "X") otherwise.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

Calculate_Thakar2015(titer_list)
}
\references{
Thakar J, et al. (2015) Aging-dependent alterations in gene expression and a mitochondrial signature of responsiveness to human influenza vaccination. Aging (Albany NY) 7(1):38-52.
}
\author{
Stefan Avey
}
