% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_TRI.R
\name{Calculate_TRI}
\alias{Calculate_TRI}
\title{Calculate TRI}
\usage{
Calculate_TRI(
  dat_list,
  subjectCol = "SubjectID",
  discretize = c(0.2, 0.3),
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"),
  na_action = "na.fail",
  ...
)
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{discretize}{a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20\% and 30\%.}

\item{responseLabels}{names for low, moderate and high responses}

\item{na_action}{how should missing \code{NA} values be treated. Default is "na.fail"}

\item{...}{Additional arguments passed to \code{lm}}
}
\value{
A list with the following elements:
\describe{
  \item{models}{the models calculated on each strain separately (with names the same as on \code{dat_list})}
  \item{residualMatrix}{the matrix of residuals}
  \item{scores}{a data frame containing the four scores (before scaling)}
  \item{TRI}{a named vector containing the continuous TRI endpoint}
  \item{TRI_d<X>}{a named vector containing the discrete TRI endpoint with cutoffs defined by the <X>\% quantile (may be more than 1, see \code{discretize})}
}
}
\description{
\code{Calculate_TRI} calculates the Titer Response Index (TRI)
}
\details{
Calculates the Titer Response Index (TRI) defined in Bucasas et al. 2011
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

## Calculate the titer response index (TRI)
endpoints <- Calculate_TRI(titer_list)
summary(endpoints)

## Get discrete endpoints using upper/lower 30\%
endpoints$TRI_d30

## Recreate Supp. Fig. S1
pairs(endpoints$scores, col = endpoints$TRI_d30)
}
\references{
Bucasas KL, et al. (2011) Early patterns of gene expression correlate with the humoral immune response to influenza vaccination in humans. J Infect Dis 203(7):921-9.
}
\seealso{
\code{lm}
}
\author{
Stefan Avey
}
