% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePairedAdjMFC.R
\name{CalculateD0NormPaired}
\alias{CalculateD0NormPaired}
\title{CalculateD0NormPaired}
\usage{
CalculateD0NormPaired(dat, fcStdCols = grep("fc_std_norm", colnames(dat),
  value = TRUE))
}
\arguments{
\item{dat}{data frame containing \code{fcStdCols}}

\item{fcStdCols}{column names containing the titer fold changes for each strain
standardized across subjects}
}
\value{
a numeric vector containing the values from d0StdCols that
                 correspond to the maximum over the strains of fcStdCols
}
\description{
\code{CalculateD0NormPaired} calculates the normalized day 0 titer paired with the titer with maximum normalized fold change
}
\details{
If there are multiple strains that have the maximal fold change,
choose the day 0 titer that is higher since this will allow
for a greater adjustment and better chance of being a high responder.

Column names containing the day 0 titers for each strain
standardized across subjects are assumed to follow the same
pattern as \code{fcStdCols} with "d0" replacing "fc" in the name.
}
\examples{
## First Example

}
\author{
Stefan Avey
}

