% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatemaxRBA.R
\name{CalculatemaxRBA}
\alias{CalculatemaxRBA}
\title{Calculate maxRBA}
\usage{
CalculatemaxRBA(dat_list, subjectCol = "SubjectID", method = c("exp", "lm"),
  yMinZero = FALSE, scoreFun = max, discretize = c(0.2, 0.3),
  normalize = TRUE, scaleResiduals = FALSE,
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"),
  na_action = "na.fail", ...)
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{method}{a character string specifying the method used to model the relationship between day 0 and fold change values. One of either "lm" for a linear model or "exp" for an exponential model.}

\item{yMinZero}{a logical specifying whether fitted y values below 0 should be set to 0.}

\item{scoreFun}{a function applied to all (potentially scaled) residuals for each subject to determine the endpoint. Default is \code{max} but \code{sum} may also be useful to quantify the total response.}

\item{discretize}{a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20\% and 30\%.}

\item{normalize}{Logical specifying whether residuals should be normalized with the inverse normal transform. Default is \code{TRUE}.}

\item{scaleResiduals}{Logical. Should residuals be scaled inversely by the
square of the confidence intervals from the linear model.}

\item{responseLabels}{names for low, moderate and high responses}

\item{na_action}{how should missing \code{NA} values be treated. Default is "na.fail"}

\item{...}{Additional arguments passed to \code{lm} if method == "lm" or \code{nls} if method == "exp"}
}
\value{
A list with the following elements:
\describe{
  \item{models}{the models calculated on each strain separately (with names the same as on \code{dat_list})}
  \item{residualMatrix}{the matrix of residuals}
  \item{maxRBA}{a list containing the continuous and discrete maxRBA metrics}
}
}
\description{
\code{CalculatemaxRBA} calculates the maximum residual after baseline-adjustment for each viral strain
}
\details{
Calculates the baseline-adjusted fold change for each strain of virus
using (unnormalized) fold change and baseline titers. Linear regression or
an exponential curve is used to remove the effect of baseline titers on fold changes.
The score function (\code{scoreFun}) is used to combine the adjusted fold change across
multiple strains.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

## Using a linear fit
endpoints <- CalculatemaxRBA(titer_list, method = "lm")
summary(endpoints)
## Get discrete endpoints using upper/lower 30\%
endpoints$maxRBA_d30

## Get endpoints with a 50\% split into high and low
endpoints <- CalculatemaxRBA(titer_list, method = "exp", discretize = 0.5)
endpoints$maxRBA_d50
}
\author{
Stefan Avey
}
\seealso{
\code{lm, nls}
}

