% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSAdjMFC.R
\name{CalculateSAdjMFC}
\alias{CalculateSAdjMFC}
\title{Calculate SAdjMFC}
\usage{
CalculateSAdjMFC(datList, subjectCol = "SubjectID", method = c("lm", "exp"),
  yMinZero = FALSE, scoreFun = max, fcCol = "fc", d0Col = "d0",
  normalize = TRUE, discretize = c(0.2, 0.3), scaleResiduals = FALSE,
  responseLabels = paste0(c("low", "moderate", "high"), "Responder"),
  na_action = "na.fail", ...)
}
\arguments{
\item{datList}{a list with one data frame for each strain and each data frame containing the columns \code{fcCol} and \code{d0Col}. The order of each data frame must be the same and they must be the same dimensions. In addition, each data frame must be sorted by \code{d0Col} from low to high.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{method}{a character string specifying the method used to model the relationship between day 0 and fold change values. One of either "lm" for a linear model or "exp" for an exponential model.}

\item{yMinZero}{a logical specifying whether fitted y values below 0 should be set to 0.}

\item{scoreFun}{a function applied to all (potentially scaled) residuals for each subject to determine the endpoint. Default is \code{max} but \code{sum} may also be useful to quantify the total response.}

\item{fcCol}{character string specifying the name of the fold change column in each element of \code{datList}}

\item{d0Col}{character string specifying the name of the day 0 column in each element of \code{datList}}

\item{normalize}{Logical specifying whether residuals should be normalized with the inverse normal transform. Default is \code{TRUE}.}

\item{discretize}{a vector of quantiles in (0, 0.5] specifying where to make the cutoff for low, moderate and high responses. Default is 20\% and 30\%.}

\item{scaleResiduals}{Logical. Should residuals be scaled inversely by the
square of the confidence intervals from the linear model.}

\item{responseLabels}{names for low, moderate and high responses}

\item{na_action}{how should missing \code{NA} values be treated. Default is "na.fail"}

\item{...}{Additional arguments passed to \code{lm} if method == "lm" or \code{nls} if method == "exp"}
}
\value{
A list with the following elements:
        "models":         the models calculated on each strain separately (with names the same as on \code{datList})
        "residualMatrix": the matrix of residuals
        "SAdjMFC":        a list containing the continuous and discrete SAdjMFC metrics
}
\description{
\code{CalculateSAdjMFC} calculates the baseline-adjusted maximum fold change (MFC)
for each viral strain
}
\details{
Calculates the baseline-adjusted fold change for each strain of virus
using (unnormalized) fold change and baseline titers. Linear regression or
an exponential curve is used to remove the effect of baseline titers on fold changes.
The score function (\code{scoreFun}) is used to combine the adjusted fold change across
multiple strains.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## First Example

}
\author{
Stefan Avey
}
\seealso{
\code{lm, nls}
}
\keyword{HIPC}

