% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatewhoResp.R
\name{CalculatewhoResp}
\alias{CalculatewhoResp}
\title{Calculate whoResp}
\usage{
CalculatewhoResp(dat_list, subjectCol = "SubjectID")
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}
}
\value{
A named vector containing the response ("NR", "X", or "R") for each subject
}
\description{
\code{CalculatewhoResp} calculates the maximum residual after baseline-adjustment for each viral strain
}
\details{
Calculates the baseline-adjusted fold change for each strain of virus
using (unnormalized) fold change and baseline titers. Linear regression or
an exponential curve is used to remove the effect of baseline titers on fold changes.
The score function (\code{scoreFun}) is used to combine the adjusted fold change across
multiple strains.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

CalculatewhoResp(titer_list)
}
\author{
Stefan Avey
}

