% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatTiters.R
\name{FormatTiters}
\alias{FormatTiters}
\title{Format antibody titers.}
\usage{
FormatTiters(titers, log2Transform = TRUE, fcMinZero = TRUE)
}
\arguments{
\item{titers}{a data frame containing one row per subject per strain. The following columns are required:
\describe{
  \item{SubjectID}{Subject IDs (column name can vary)}
  \item{Strain}{The name of the viral strain for the observation}
  \item{Pre}{The pre-vaccination (or pre-infection) titer}
  \item{Post}{The post-vaccination (or post-infection) titer}
  \item{...}{Other columns which will be preserved}
}}

\item{log2Transform}{logical specifying whether titer values should be log2 transformed}

\item{fcMinZero}{should negative fold changes be set to 0? Default is \code{TRUE}}
}
\value{
a list of data frames with one data frame per viral strain containing the "Pre" and "Post" titer measurements (row names are removed).
}
\description{
\code{FormatTiters} formats titers into a list with one tidy data frame per viral strain
}
\examples{
titer_list <- FormatTiters(Year1_Titers, log2Transform = TRUE, fcMinZero = TRUE)
}
\author{
Stefan Avey
}

