% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{GetEqn}
\alias{GetEqn}
\title{Get Formatted Model Equation}
\usage{
GetEqn(m)
}
\arguments{
\item{m}{a model object}
}
\description{
\code{GetEqn} gets the equation for various models in a human readable format
}
\examples{
## First Example

}
\author{
Stefan Avey
}
\references{
original lm_eqn and inspiration from this SO post \url{http://stackoverflow.com/questions/7549694/ggplot2-adding-regression-line-equation-and-r2-on-graph}.
}
\keyword{aveytoolkit}

