% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Multiplot}
\alias{Multiplot}
\title{Multiple ggplot2 plots on the same page}
\usage{
Multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\description{
Multiple Plot Function for ggplot
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right,
and 3 will go all the way across the bottom.
}
\author{
R Cookbook
}
\references{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_\%28ggplot2\%29/}
}

