% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Year1_Titers.R
\docType{data}
\name{Year1_Titers}
\alias{Year1_Titers}
\title{Year 1 titers.}
\format{A data frame with 42 rows and 11 variables:
\describe{
  \item{SubjectID}{a unique subject identifier}
  \item{AgeGroup}{age of subject. 20-35 (Young),  65+ (Older)}
  \item{Strain}{The name of the viral strain for the observation}
  \item{Pre}{The pre-vaccination titer}
  \item{Post}{The post-vaccination titer}
}}
\usage{
Year1_Titers
}
\description{
Antibody titers to 3 strains of influenza in a cohort of young and older adults from Yale during the 2010-2011 flu season.
}
\references{
Thakar J, et al. (2015) Aging-dependent alterations in gene expression and a mitochondrial signature of responsiveness to human influenza vaccination. Aging (Albany NY) 7(1):38-52. \url{https://www.ncbi.nlm.nih.gov/pubmed/25596819}
}
\keyword{datasets}

