% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleChart.R
\name{BubbleChart}
\alias{BubbleChart}
\title{Bubble Chart}
\usage{
BubbleChart(dat_list, fit = NULL, subjectCol = "SubjectID",
  xlimits = c(1.5, 10.5), xbreaks = 2:10, plot = TRUE, cols = 2)
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}}

\item{fit}{what type of fit to add. Current options are "lm" for linear model, "exp" for exponential, or \code{NULL} for no smoothing.}

\item{xlimits}{the x-axis limits (passed to \code{scale_x_continuous})}

\item{xbreaks}{the x-axis breaks (passed to \code{scale_x_continuous})}

\item{plot}{logical indicating whether to plot or not. Default is TRUE}

\item{cols}{numeric specifying how many columns to layout plot}

\item{scale_y}{a character string specifying whether the y axis should be "fixed" for all strains or "free".}
}
\value{
(invisibly) a list of ggplot2 objects.
}
\description{
\code{BubbleChart} visualizes baseline vs fold change in titers
}
\details{
This plot was designed for HAI titer data with baseline columns and fold change columns for multiple strains.
}
\examples{
## Prepare the data
library(dplyr)
library(ggplot2)
titers <- filter(Year2_Titers, AgeGroup == "Young")
strains <- c("A_California_7_2009", "A_Perth_16_2009", "B_Brisbane_60_2008")
titer_list <- FormatTiters(titers, strains, subjectCol = "YaleID")

## Basic plot without any fitted model
BubbleChart(titer_list)

## Change layout to plot 3 strains in a single column
BubbleChart(titer_list, cols = 1)

## Add a linear fit
BubbleChart(titer_list, fit = "lm", subjectCol = "YaleID")

## Add an exponential fit
BubbleChart(titer_list, fit = "exp", subjectCol = "YaleID")
}
\author{
Stefan Avey
}
\seealso{
\code{FormatTiters}
}
\keyword{HIPC}

