% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubbleChart.R
\name{BubbleChart}
\alias{BubbleChart}
\title{Bubble Chart}
\usage{
BubbleChart(dat_list, fit = NULL, yMinZero = FALSE,
  eqSize = 6/log2(length(dat_list) + 1), subjectCol = "SubjectID",
  colorBy = NULL, xlimits = c(1.5, 10.5), xbreaks = 2:10,
  ylimits = c(-0.5, 10), ybreaks = seq(0, 10, 2), plot = TRUE, cols = 2,
  ...)
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}. Values are assumed to be log2-transformed.}

\item{fit}{what type of fit to add. Current options are "lm" for linear model, "exp" for exponential, or \code{NULL} for no smoothing.}

\item{yMinZero}{a logical specifying whether fitted y values below 0 should be set to 0.}

\item{eqSize}{Text size of the equation. Only relevant if \code{fit} is not \code{NULL}}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{colorBy}{a character string specifying an endpoint to colorBy or \code{NULL} (default) for no coloring.}

\item{xlimits}{the x-axis limits (passed to \code{scale_x_continuous})}

\item{xbreaks}{the x-axis breaks (passed to \code{scale_x_continuous})}

\item{ylimits}{the y-axis limits (passed to \code{scale_y_continuous})}

\item{ybreaks}{the y-axis breaks (passed to \code{scale_y_continuous})}

\item{plot}{logical indicating whether to plot or not. Default is TRUE}

\item{cols}{numeric specifying how many columns to layout plot}

\item{...}{other arguments besides \code{method} and \code{subjectCol} passed to \code{\link{CalculateSAdjMFC}}.}
}
\value{
(invisibly) a list of ggplot2 objects.
}
\description{
\code{BubbleChart} visualizes baseline vs fold change in titers
}
\details{
This plot was designed for HAI titer data with baseline columns and fold change columns for multiple strains.
}
\examples{
## Prepare the data
strains <- c("A_California_7_2009", "A_Perth_16_2009", "B_Brisbane_60_2008")
titer_list <- FormatTiters(Year2_Titers, strains, subjectCol = "YaleID",
                           otherCols = "AgeGroup")

## Basic plot without any fitted model
BubbleChart(titer_list)

## Change layout to plot 3 strains in a single column
BubbleChart(titer_list, cols = 1)

## Add a linear fit
BubbleChart(titer_list, fit = "lm", subjectCol = "YaleID")

## Add an exponential fit
BubbleChart(titer_list, fit = "exp", subjectCol = "YaleID")

## Add coloring by age
BubbleChart(titer_list, fit = "exp", subjectCol = "YaleID", colorBy = "AgeGroup")
}
\author{
Stefan Avey
}
\seealso{
\code{FormatTiters}
}
\keyword{HIPC}

