% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatewhoResp.R
\name{CalculatewhoResp}
\alias{CalculatewhoResp}
\title{Calculate whoResp}
\usage{
CalculatewhoResp(dat_list, subjectCol = "SubjectID")
}
\arguments{
\item{dat_list}{a named list like the one returned by \code{\link{FormatTiters}}.}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}
}
\value{
A named list with 1 element named "whoResp" containing the response ("NR", "X", or "R").
}
\description{
\code{CalculatewhoResp} calculates a response definition similar to the WHO defintion using a 4-fold cutoff.
}
\details{
Subjects are responders ("R") if they acheive a 4-fold or greater fold change in
titer to at least 2 strains, nonresponders ("NR") if they do not acheive a 4-fold
or greater fold change in titer to any strain, and intermediate ("X") otherwise.
Missing (\code{NA}) values are handled by being returned as missing in the
endpoints in the output
}
\examples{
## Prepare the data
titer_list <- FormatTiters(Year2_Titers)

CalculatewhoResp(titer_list)
}
\author{
Stefan Avey
}

