% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatTiters.R
\name{FormatTiters}
\alias{FormatTiters}
\title{Format antibody titers.}
\usage{
FormatTiters(titers, strains, subjectCol = "SubjectID",
  otherCols = vector(mode = "character"), d0Cols = paste0("d0_", strains),
  fcCols = paste0("fc_", strains), fcMinZero = TRUE, log2Transform = TRUE)
}
\arguments{
\item{titers}{a data frame containing the titer information}

\item{strains}{the names of the virus strains}

\item{subjectCol}{the name of the column specifying a subject ID. Default is "SubjectID".}

\item{otherCols}{a character vector specifying which additional columns of titers to retain. (Defaults to an empty character vector).}

\item{d0Cols}{the column names of day 0 (baseline) columns}

\item{fcCols}{the column names of fold change columns}

\item{fcMinZero}{should negative fold changes be set to 0? Default is \code{TRUE}}

\item{log2Transform}{logical specifying whether titer values should be log2 transformed}
}
\value{
a list of data frames with one data frame per viral strain containing the baseline ("d0"), fold change ("fc") and any other columns specified by the \code{otherColumns} argument.
}
\description{
\code{FormatTiters} formats titers into a list with one tidy data frame per viral strain
}
\examples{
\dontrun{
## Example using the master phenotype file
library(dplyr)
titers <- master \%>\%
  filter(Year == 1, AgeGroup \%in\% "Young", !is.na(whoResp))
titer_list <- FormatTiters(titers,
                           strains = c("A_California_7_2009",
                               "A_Perth_16_2009",
                               "B_Brisbane_60_2008"))
}
}
\author{
Stefan Avey
}
\keyword{HIPC}

