% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Multiplot}
\alias{Multiplot}
\title{Multiple ggplot2 plots on the same page}
\usage{
Multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\description{
Multiple Plot Function for ggplot
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right,
and 3 will go all the way across the bottom.
}
\examples{
library(ggplot2)

## This example uses the ChickWeight dataset, which comes with ggplot2
## First plot
p1 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet, group=Chick)) +
  geom_line() +
  ggtitle("Growth curve for individual chicks")

                                        # Second plot
p2 <- ggplot(ChickWeight, aes(x=Time, y=weight, colour=Diet)) +
  geom_point(alpha=.3) +
  geom_smooth(alpha=.2, size=1) +
  ggtitle("Fitted growth curve per diet")

                                        # Third plot
p3 <- ggplot(subset(ChickWeight, Time==21), aes(x=weight, colour=Diet)) +
  geom_density() +
  ggtitle("Final weight, by diet")

                                        # Fourth plot
p4 <- ggplot(subset(ChickWeight, Time==21), aes(x=weight, fill=Diet)) +
  geom_histogram(colour="black", binwidth=50) +
  facet_grid(Diet ~ .) +
  ggtitle("Final weight, by diet") +
  theme(legend.position="none")        # No legend (redundant in this graph)

Multiplot(p1, p2, p3, p4, cols=2)
}
\author{
R Cookbook
}
\references{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_\%28ggplot2\%29/}
}

